Refactor WeekPlanPage component by modularizing sections into dedicated components for better organization and maintainability. Removed unused dark mode styles from globals.css. Enhanced code readability and separation of concerns.
