+++ components/ui/input.tsx
@@ -1,0 +1,17 @@
+import { InputHTMLAttributes, forwardRef } from "react";
+
+type InputProps = InputHTMLAttributes<HTMLInputElement> & {
+  className?: string;
+};
+
+export const Input = forwardRef<HTMLInputElement, InputProps>(
+  ({ className = "", ...props }, ref) => {
+    const base =
+      "w-full px-3 py-2 border border-gray-300 rounded-md focus:outline-none focus:ring-2 focus:ring-blue-500";
+    return <input ref={ref} className={`${base} ${className}`} {...props} />;
+  }
+);
+
+Input.displayName = "Input";
+
+
