+++ components/ui/textarea.tsx
@@ -1,0 +1,17 @@
+import { TextareaHTMLAttributes, forwardRef } from "react";
+
+type TextareaProps = TextareaHTMLAttributes<HTMLTextAreaElement> & {
+  className?: string;
+};
+
+export const Textarea = forwardRef<HTMLTextAreaElement, TextareaProps>(
+  ({ className = "", ...props }, ref) => {
+    const base =
+      "w-full px-3 py-2 border border-gray-300 rounded-md focus:outline-none focus:ring-2 focus:ring-blue-500";
+    return <textarea ref={ref} className={`${base} ${className}`} {...props} />;
+  }
+);
+
+Textarea.displayName = "Textarea";
+
+
